Import Spring Boot's dependency BOM, fix spring-boot:run at parent project level (#276)
* Fix using spring-boot:run from project root

A little fix when some subproject modules are Spring Boot apps, but the
parent is not.

Also document how to overcome IntelliJ CE woes with Spring Boot, and
inter-module Maven dependencies.

* Get Spring managed dep versions via their BOM

Spring has done the work of finding direct and transitive versions of
dependencies compatible with their ecosystem, and declaring them as a
BOM. It's in our best interest to heed it as much as we can.

* ci: Try --also-make instead of skipping Enforcer

Per discussion:

  https://github.com/gojek/feast/pull/271#discussion_r339412421

The `install` step should not be necessary for the unit tests, but we
should use `--batch-mode` on the test step for CI.

* Remove maven-shade-plugin from core

It was resulting in a non-functional JAR, appears to be incomplete
configuration. It seems unlikely to me that core will be used as a
library, if there is a need the shading configuration can be fixed but
it probably needs other considerations anyway.