Backport Sonatype publishing and datatypes module for v0.3.x (#407)
* Use Nexus staging plugin for deployment (#394)

* Use Nexus staging pluging for deployment

* Fix Javadoc error

* Hard coded parent version as variable substitution is not supported

* Introduce datatypes/java module for proto generation

Rather than the Maven protobuf plugin running on the same symlinked
definitions in several Java modules, localize this process into one
module that the others depend on.

This provides a single module that can be depended on by third-party
extensions with the bare minimum of dependencies.

Also removes proto files that are no longer used.

* Java SDK release script (#406)

* Use back revision variable in pom.xml
So user or CI system can easily override revision from external sources such as Git tag name

* Add flatten maven plugin
This plugin is useful during deployment so the final pom is resolved without parent dependency, i.e. we do not necessarily need to upload parent library

* Increase versions for maven source,javadoc,spotless plugins
So it has newer features and more fixes

* Add gpg-plugin needed to sign releases

* Use oss configure for flatten plugin, add developers info in pom.xml (required for releasing library

* Add publish-java-sdk script

* Add more logs to publish-java-sdk.sh

* Add ProwJob publish-java-sdk

* Use GPG_KEY_IMPORT_DIR variable

* Update revision in pom.xml to 0.4.2-SNAPSHOT

* Publish datatypes/java along with sdk/java

Co-authored-by: Khor Shu Heng <32997938+khorshuheng@users.noreply.github.com>
Co-authored-by: David Heryanto <david.heryanto@hotmail.com>