Enforce JDK 11 for development, but build Ingestion to target Java 8 (#518)
* Require Java 11 with Enforcer

We build for Java 11 now, so the build will fail with older JDKs. Have
the Enforcer plugin do that for a more lucid error message.

This reverts 190e605 which was squash-merged in a larger commit.

Partially addresses #517

* Build Ingestion to target Java 8, for Beam compat

As well as datatypes-java since ingestion depends on it.

Java 11 is desirable for the other components, but for Beam it may
impose limitations on what runners Feast can support, if it is even safe
to run on an 11 JRE now.

https://issues.apache.org/jira/browse/BEAM-2530

References #517