Refactor sendMultiGetAndProcess() to getFeaturesForFeatureSet().
Rename as the original name described the implementation instead what
the goal of the method is.

Construct and pass the decoder outside method in getFeaturesForFeatureSet()
instead of inside method in sendMultiGetAndProcess() as constructing it
inside added uncessary arguments and code (ie getting the string ref
out a redis key seemed out of place.)

Changed getFeaturesForFeatureSet() to return null instead of a nullFeatureRow.
when retrieving data from redis using sendMultiGet() using a key missing
in Redis. This makes the missing key condition easier to detect.

Removed unncessary arguments passed to sendMultiGetAndProcess()
- featureReference used to contruct the nullFeatureRow
- featureSetSpec used to contruct the decoder.