Fix config validation for feast.jobs.metrics.host (#662)
* Update config validation for feast.jobs.metrics.host

`host` will be used as the parameter to connect to a metrics server e.g StatsD Server. It's not a URL but rather an IP address or hostname. For example StatsD UDP server expects a DNS hostname or IP address and not a URL such as http://10.23.40.1. The UDP server does not understand http application layer protocol.

* Typo in error message