JobUpdateTask cleanups (#650)
* core: Clean up repetition in JobUpdateTask

Logic is a bit complicated here and trimming excess noise starts to help
make it quicker to grok.

We were doing the equivalent of Set#equals by hand.

* core: Reduce noise of AuditLogger in JobUpdateTask

* core: Simplify JobUpdateTask internal helper calls

The helper methods like startJob and updateJob had parameters that are
instance state, so there seemed to be no reason to pass around arguments
for them.

* core: Make isTerminal an instance property of JobStatus

* core: Remove test setup duplication in JobUpdateTaskTest

* core: Add unit tests for JobStatus

* core: Refactor JobUpdateTask to use domain model types

And factor out things that don't differ between its test cases.

JobCoordinatorService unmarshals protos to model types eagerly as they
come off the wire, so we're left dealing with models everywhere else.