Run end-2-end tests in GitHub Actions CI (#6)
Uses docker-compose to deploy alongside with Feast (Core, Kafka, PostgreSQL) containers, a CI container that runs end-to-end tests.

As docker-compose runs the container in the same network, it can access the other containers directly through the docker network.

With this PR we run a single basic integration test for Feast Core only, to validate the concept. Later on we will try to also deploy the serving container, and run more end-to-end tests.