docs: Move release process updates into its new dedicated page
On master we still have a monolithic `docs/contributing/contributing.md`
_and_ redundant copies of some of the information in broken out pages
(that have missed some useful updates to the monolith). Needs to get
sorted out, but going ahead and moving release process to the dedicated
page that now exists for it, in part to port easily to v0.4 docs where
the monolith is gone.