Make storage modules target Java 8, since Ingestion must
With the addition of the Storage APIs and connector implementations, we
did not specify that their build should target Java 8 bytecode. Because
Ingestion depends on them, we regressed on keeping Ingestion compatible
with Java 8 runtimes according to #518.

This could be confirmed prior to this change with, for example:

    $ mvn -pl ingestion clean compile
    $ javap -v storage/connectors/redis/target/classes/feast/storage/connectors/redis/writer/RedisCustomIO.class | grep major
      major version: 55

I take this as proof that a blacklist approach with 11 as the default in
our root POM is error prone, so this change flips it so that 8 is
default and our leaf applications that are able use 11 opt into it in
whitelist fashion.