Fix Docker Compose Batch Serving (#744)
* Fix issue where external Feast client cannot access docker-compose Kafka

This issue occurs when users attempt to interact with an internal
docker-compose Kafka from a extenral. When Feast Core provides the
Kakfa bootstrap servers, its relative to inside the docker-compose
network, and not resolvable from the outside. This adds an
external alias for the internal Kafka to bootstrap servers
so that external clients work correctly.

* Fix batch serving listening on wrong port and incorrect active-store typo

* Fix extra slash in gcs store url.