CI/CD pipeline on SparkRunner (#36)
Closes KE-687, KE-688

Deploys CI/CD pipeline on SparkRunner, and run tests for Online serving.

Developed a smaller representative subset of tests, to limit resource consumption on Databricks.

NB: The Spark job also produces output Delta files in Azure Storage. That functionality is currently not integration tested, as that will be achieved with the Historical Serving feature.

NB: an Online Retriever must be connected for Feast to be spun up, but it's implemented as null until we integrate the real retriever.