Apply labels when creating Dataflow jobs.
The job labels are derived from the labels of the FeatureSets that
the job is expected to process when the Dataflow job is created.

When the Dataflow job is processing multiple FeatureSets and there
are 2 FeatureSets with the same label key but different label value,
the resulting label value for the job will be the concatentation
of each label value from the FeatureSets.