Initializing `ttl` and `ttl_duration` to None
`ttl` is noted as an optional input in the `FeatureView` class which should default to `None`. If the FeatureView's `ttl` input parameter is None (by default or specified by the User when creating the feature view), the `ttl_duration` is never initialized inside the `FeatureView.to_proto()` function and will produce an `UnboundLocalError` since the local variable `ttl_duration` is referenced before assignment.