Improve performance of _convert_arrow_to_proto
_convert_arrow_to_proto does multiple lookups in the list of column
names for the arrow table to find the index of columns. This is done for
each row so the time complexity grows quickly.
Rather than search the list of names for the index we construct a
dictionary {column_name: index}. This allows faster lookups for the
index of a column and speeds up the method significantly which affects
e.g. materialization.

Signed-off-by: Gunnar Sv Sigurbjörnsson <gunnar.sigurbjornsson@gmail.com>