Add integration tests for AWS Lambda feature server (#2001)
* Add integration tests for AWS Lambda feature server

Signed-off-by: Tsotne Tabidze <tsotne@tecton.ai>

* Fix feature server bugs & integration test

Signed-off-by: Tsotne Tabidze <tsotne@tecton.ai>

* Address comments & ignore incorrect linter errors

Signed-off-by: Tsotne Tabidze <tsotne@tecton.ai>

* Fix float comparison & lambda api call retries

Signed-off-by: Tsotne Tabidze <tsotne@tecton.ai>

* Increase retries and catch other ClientErrors

Signed-off-by: Tsotne Tabidze <tsotne@tecton.ai>

* Fix lambda name shortening (md5 hashing instead of base64 prefixing)

base64-ing project name and then taking the prefix was not working when
the prefix of the project name was identical between tests (only their
endings were different for some tests). This caused lambda names between
test cases to be the same, causing conflicts between updates. Now, we
instead use md5 hash which depends on the entire body of the input,
eliminating this issue.

Signed-off-by: Tsotne Tabidze <tsotne@tecton.ai>