Don't allow FeatureStore.apply with commit=False
The apply method will call update_infra for the provider without
offering a dry run, so invoking it without committing should not be
an option. It would make all the infra updates but not update the
registry causing them to be out of sync.

Added commit=False to apply_feature_service and apply_feature_table in the
apply method.
Today when calling apply, the update_infra might fail but then we have
already updated the registry causing everything to be out of sync.

Signed-off-by: Gunnar Sv Sigurbjörnsson <gunnar.sigurbjornsson@gmail.com>