docs: Install instructions fix directory issue (#3290)
What this PR does / why we need it:
When you run 'feast init my_feature_repo', it makes a 'my_feature_repo' directory. However the 'feature_store.yaml' file is within another directory called 'feature_repo'. So the feast apply command gives an error if you are simply in the 'my_feature_store' directory.

Which issue(s) this PR fixes:
This fixes a minor issue in the README with the install instructions

Fixes #

Signed-off-by: Danny Chiao <danny@tecton.ai>

Signed-off-by: Danny Chiao <danny@tecton.ai>