chore: Remove upper bound from fsspec requirement
In #3922 the fsspec requirement was loosened to allow what was then
more recent releases.  Since then Feast has further reduced the
number dependencies with upper bounds.  Instead of re-testing each
month (when fsspec usually releases) it would now be idiomatic only
specify a minimum bound.

The new minimum bound was chosen to match what was last locked in ci,
since that is what was tested and we can be less sure older versions
also work.

Signed-off-by: Chris Burroughs <chris.burroughs@gmail.com>