chore: Define Node.js version with Node Version Manager in Feast UI
Defining the Node.js version in a single file makes it easier to use a
consistent version and upgrade it in the future. v20.18.0 is the current
latest LTS version. Note that the @types/node version numbers don't match
the related Node.js version numbers, 20.16.13 is the latest package in the
major number 20.

setup-node GitHub action node-version-file documentation:
https://github.com/actions/setup-node/blob/main/docs/advanced-usage.md#node-version-file

Signed-off-by: Harri Lehtola <peruukki@hotmail.com>