chore: Disable strict mode warnings in scripts and config in /ui
Strict mode is implicit in modules, so ESLint gave warnings about using
them in the scripts. However, the scripts are run directly with `node`, so
they are not used as modules. But setting their sourceType to "script"
results in ESLint errors saying "Strict mode is not permitted." I couldn't
figure out why, so let's just disable the rule for the scripts.

Later I noticed the same warning in the config directory, so let's include
that too.

Signed-off-by: Harri Lehtola <peruukki@hotmail.com>