chore: Eject Create React App dependency from Feast UI (#4681)
* chore: Eject create-react-app with `yarn eject` in /ui

Signed-off-by: Harri Lehtola <peruukki@hotmail.com>

* chore: Move most eject dependencies to devDependencies in /ui

Signed-off-by: Harri Lehtola <peruukki@hotmail.com>

* chore: Remove duplicate babel config added when ejecting in /ui

The presets are set in .babelrc.js.

Signed-off-by: Harri Lehtola <peruukki@hotmail.com>

* chore: Upgrade dependencies with vulnerabilities in /ui

These two dependencies can now be upgraded after ejecting.

Signed-off-by: Harri Lehtola <peruukki@hotmail.com>

* chore: Extract Jest configuration from package.json in /ui

This keeps package.json cleaner.

Signed-off-by: Harri Lehtola <peruukki@hotmail.com>

* chore: Extract ESLint configuration from package.json in /ui

Signed-off-by: Harri Lehtola <peruukki@hotmail.com>

* chore: Remove unused tailwindcss added when ejecting in /ui

Signed-off-by: Harri Lehtola <peruukki@hotmail.com>

* chore: Disable strict mode warnings in scripts and config in /ui

Strict mode is implicit in modules, so ESLint gave warnings about using
them in the scripts. However, the scripts are run directly with `node`, so
they are not used as modules. But setting their sourceType to "script"
results in ESLint errors saying "Strict mode is not permitted." I couldn't
figure out why, so let's just disable the rule for the scripts.

Later I noticed the same warning in the config directory, so let's include
that too.

Signed-off-by: Harri Lehtola <peruukki@hotmail.com>

---------

Signed-off-by: Harri Lehtola <peruukki@hotmail.com>