chore: Enable one future react-router-dom flag
react-router-dom was writing warning messages in the console about
upcoming features in v7 and suggested opting in early. I tried turning
them on, but couldn't get custom tabs navigation working properly with
v7_relativeSplatPath enabled, so decided to keep it off for now. (Plus
it might also break something in consuming apps.)

v7_startTransition only needed an update in one place in a test where we
need to await for the UI to update, so kept it on. It also shouldn't
require any changes in consuming application code, though possibly
something in tests like we needed here.

Related docs:

- https://reactrouter.com/en/6.28.0/upgrading/future#v7_relativesplatpath

- https://reactrouter.com/en/6.28.0/upgrading/future#v7_starttransition

Signed-off-by: Harri Lehtola <peruukki@hotmail.com>