chore: Convert EuiCustomLink to TypeScript in /ui
This was the last JavaScript component, convert it to TypeScript too.

The `href` prop was never used, so we omit it from the props type
definition, and no longer pass it to the component.

We could also have chosen to drop the `to` prop and use `href` instead,
but `to` is used by React Router's Link
(https://reactrouter.com/6.29.0/components/link), and it makes sense to
use the same name in EuiCustomLink since we pass it to React Router's
`navigate` function; the naming hints that it's tied to React Router and
not just a regular URL path.

Signed-off-by: Harri Lehtola <peruukki@hotmail.com>