feat!: Include PUBLIC_URL in defaultProjectListPromise URL in /ui
We currently always fetch the project list from the root path by
default, even if the UI is served from a non-root path via PUBLIC_URL.

It seems reasonable to assume that the project list would be served
from the same path as the UI by default, so change the default project
list URL to include the basename from PUBLIC_URL. This way you don't
need to specify a custom `projectListPromise` for this base case, as
shown by the changes in ui/src/index.tsx.

BREAKING CHANGE:
The PUBLIC_URL environment variable is now taken into account by default
when fetching the projects list. This is a breaking change only if all
these points apply:

1. You're using Feast UI as a module

2. You're serving the UI files from a non-root path via the PUBLIC_URL
   environment variable

3. You're serving the project list from the root path

4. You're not passing the `feastUIConfigs.projectListPromise` prop to
   the FeastUI component

In this case, you need to explicitly fetch the project list from the
root path via the `feastUIConfigs.projectListPromise` prop:

```diff
 const root = createRoot(document.getElementById("root")!);
 root.render(
   <React.StrictMode>
-    <FeastUI />
+    <FeastUI
+      feastUIConfigs={{
+        projectListPromise: fetch("/projects-list.json", {
+            headers: {
+              "Content-Type": "application/json",
+            },
+          }).then((res) => res.json())
+      }}
+    />
   </React.StrictMode>
 );
```

Signed-off-by: Harri Lehtola <peruukki@hotmail.com>